refactor(archiver): move orphan-block prune out of L1 synchronizer
`pruneOrphanProposedBlocks` is wall-clock based and does not touch L1, so
it belongs on the `Archiver` rather than the `ArchiverL1Synchronizer`.
Moves the method (and its `epochCache` / `dateProvider` dependencies)
onto the `Archiver`, called directly from `sync()` between
`processInboundQueue()` and `syncFromL1()`. The synchronizer keeps the
L1-block-driven `pruneUncheckpointedBlocks` (used to clear late stale
blocks once L1 advances past their slot); its inline emit is now
duplicated in both prune paths to keep them self-contained.

No behavior change — verified by the existing orphan-prune unit tests in
`archiver-sync.test.ts` and the full archiver suite.
