+++ yarn-project/archiver/src/archiver-store.test.ts
@@ -12,6 +12,7 @@ import {
 import { Buffer32 } from '@aztec/foundation/buffer';
 import { Fr } from '@aztec/foundation/curves/bn254';
 import { EthAddress } from '@aztec/foundation/eth-address';
+import { DateProvider } from '@aztec/foundation/timer';
 import { openTmpStore } from '@aztec/kv-store/lmdb-v2';
 import { L2Block } from '@aztec/stdlib/block';
 import type { L1RollupConstants } from '@aztec/stdlib/epoch-helpers';
@@ -63,6 +64,7 @@ describe('Archiver Store', () => {
     blobClient = mock<BlobClientInterface>();
     epochCache = mock<EpochCache>();
     epochCache.getCommitteeForEpoch.mockResolvedValue({ committee: [] as EthAddress[] } as EpochCommitteeInfo);
+    epochCache.pipeliningOffset.mockReturnValue(0);
 
     const rollupContract = mock<RollupContract>();
     Object.defineProperty(rollupContract, 'address', { value: rollupAddress.toString(), writable: true });
@@ -121,6 +123,8 @@ describe('Archiver Store', () => {
       initialHeader,
       initialBlockHash,
       l2TipsCache,
+      epochCache,
+      new DateProvider(),
     );
   });
