+++ yarn-project/archiver/src/factory.ts
@@ -181,6 +181,8 @@ export async function createArchiver(
     initialHeader,
     initialBlockHash,
     l2TipsCache,
+    epochCache,
+    deps.dateProvider ?? new DateProvider(),
   );
 
   await archiver.start(opts.blockUntilSync);
